web-api/requirements.txt
absl-py==0.6.1
aniso8601==4.0.1
astor==0.7.1
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
blinker==1.4
boto3==1.7.47
botocore==1.10.47
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
codeclimate-test-reporter==0.2.3
colorama==0.4.1
connexion==2.1.0
coverage==4.3.4
cycler==0.10.0
Cython==0.29.2
decorator==4.3.0
docker==3.6.0
docker-pycreds==0.4.0
docutils==0.14
ez-setup==0.9
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
future==0.17.1
gast==0.2.0
grpcio==1.17.0
h2o==3.24.0.1
h5py==2.8.0
idna==2.7
inflection==0.3.1
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mock==2.0.0
more-itertools==4.2.0
numpy==1.16.2
openapi-spec-validator==0.2.4
paho-mqtt==1.4.0
pandas==0.23.1
parso==0.3.1
pathlib2==2.3.3
pbr==5.1.3
pd==0.0.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2-binary>=2.7.4
ptyprocess==0.6.0
py==1.5.3
py4j==0.10.7
pyarrow==0.13.0
Pygments==2.3.0
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.20.3
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.3.3
stevedore==1.30.1
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.4.3
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
